Update IEP `wallet_addEthereumChain` test
- Added sameChainID helper
- Fixed feature flag check
- Moved validation and types to provider bridge utils
